# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for the NewsBalancer application which analyzes political bias in news articles using LLM models",
        "title": "NewsBalancer API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "NewsBalancer Support",
            "url": "https://github.com/alexandru-savinov/BalancedNewsGo",
            "email": "support@newsbalancer.example"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/api/articles": {
            "get": {
                "description": "Fetches a list of articles with optional filtering by source, leaning, and pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "Get articles",
                "operationId": "getArticlesList",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by news source",
                        "name": "source",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by political leaning (left/center/right)",
                        "name": "leaning",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of articles",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/api.ArticleResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new article with the provided information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "Create article",
                "operationId": "createArticle",
                "parameters": [
                    {
                        "description": "Article information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.CreateArticleResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Article URL already exists",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ArticleResponse": {
            "type": "object",
            "properties": {
                "article_id": {
                    "type": "integer"
                },
                "composite_score": {
                    "type": "number"
                },
                "confidence": {
                    "type": "number"
                },
                "content": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "score_source": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "api.CreateArticleRequest": {
            "description": "Request body for creating a new article",
            "type": "object",
            "required": [
                "content",
                "pub_date",
                "source",
                "title",
                "url"
            ],
            "properties": {
                "content": {
                    "description": "Article content",
                    "type": "string",
                    "example": "Article content..."
                },
                "pub_date": {
                    "description": "Publication date in RFC3339 format",
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                },
                "source": {
                    "description": "News source name",
                    "type": "string",
                    "example": "CNN"
                },
                "title": {
                    "description": "Article title",
                    "type": "string",
                    "example": "Breaking News"
                },
                "url": {
                    "description": "Article URL",
                    "type": "string",
                    "example": "https://example.com/article"
                }
            }
        },
        "api.CreateArticleResponse": {
            "description": "Response from creating a new article",
            "type": "object",
            "properties": {
                "article_id": {
                    "description": "ID of the created article",
                    "type": "integer",
                    "example": 42
                },
                "status": {
                    "description": "Status of the operation",
                    "type": "string",
                    "example": "created"
                }
            }
        },
        "api.ErrorDetail": {
            "description": "Detailed error information",
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code",
                    "type": "string",
                    "example": "validation_error"
                },
                "message": {
                    "description": "Human-readable error message",
                    "type": "string",
                    "example": "Invalid input parameters"
                }
            }
        },
        "api.ErrorResponse": {
            "description": "Standard API error response",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error details",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.ErrorDetail"
                        }
                    ]
                },
                "success": {
                    "description": "Always false for errors",
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "api.StandardResponse": {
            "description": "Standard API success response",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response data payload"
                },
                "success": {
                    "description": "Always true for success",
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
...
}